1.修改sso前缀错误,图片显示正常;3.完成添加房源;2.完成房源分页展示,Hutool的StrUtil.toUnderLineCase()很好的处理了下划线和驼峰式的转换

+++ src/main/java/com/oywy/entity/HouseDetail.java
@@ -6,7 +6,7 @@ public class HouseDetail {
   private String layoutDesc;
   private String traffic;
   private String roundService;
-  private Long rentWay;
+    private Integer rentWay;
   private String address;
   private Long subwayLineId;
   private String subwayLineName;
@@ -54,11 +54,11 @@ public class HouseDetail {
     this.roundService = roundService;
   }
 
-  public Long getRentWay() {
+    public Integer getRentWay() {
     return rentWay;
   }
 
-  public void setRentWay(Long rentWay) {
+    public void setRentWay(Integer rentWay) {
     this.rentWay = rentWay;
   }
